+++ src/targetdb/__init__.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python
 
-from .manage import create_schema
-from .utils import draw_diagram, generate_schema_markdown, read_conf
+# from .manage import create_schema
+
+# from .utils import draw_diagram, generate_schema_markdown, read_conf
 
 __version__ = "0.1.0"
 
-__all__ = ["create_schema", "draw_diagram", "generate_schema_markdown", "read_conf"]
+# __all__ = ["create_schema", "draw_diagram", "generate_schema_markdown", "read_conf"]
