+++ src/targetdb/cli/pfs_targetdb_create_schema.py
@@ -3,7 +3,7 @@
 import argparse
 import sys
 
-from .. import create_schema
+from ..utils import create_schema
 
 
 def get_arguments():
