+++ src/targetdb/cli/pfs_targetdb_draw_diagram.py
@@ -6,7 +6,7 @@ import argparse
 # from logzero import logger
 from loguru import logger
 
-from .. import draw_diagram
+from ..utils import draw_diagram
 
 
 def get_arguments():
